fix(logging): keep GUI stdout log channel
