+++ package.json
@@ -20,7 +20,7 @@
     "test:plan-management-delete": "node scripts/tests/test-plan-management-delete-scope.mjs",
     "test:scheduler-domain": "npm run build && node scripts/tests/test-scheduler-domain.mjs",
     "test:daily-sortie-stats": "node scripts/tests/test-daily-sortie-stats.mjs",
-    "test:logger-dedup": "node scripts/tests/test-logger-dedup.mjs",
+    "test:gui-log-route": "node scripts/tests/test-gui-log-route.mjs",
     "test:ship-library-updater": "node scripts/tests/test-ship-library-updater.js",
     "test:ocr-log-analyzer": "node scripts/tests/test-ocr-log-analyzer.js",
     "test:map-intel": "node scripts/tests/test-map-intel.js",
